Fix `kernel_size` and `sigma` in gaussian_blur.
