+++ torchvision/prototype/transforms/functional/_misc.py
@@ -1,4 +1,4 @@
-from typing import List, Optional
+from typing import List, Union
 
 import PIL.Image
 import torch
@@ -21,7 +21,7 @@ def normalize(
 
 
 def gaussian_blur_image_tensor(
-    img: torch.Tensor, kernel_size: List[int], sigma: Optional[List[float]] = None
+    img: torch.Tensor, kernel_size: Union[int, List[int]], sigma: Union[int, float, List[float], None] = None
 ) -> torch.Tensor:
     if isinstance(kernel_size, int):
         kernel_size = [kernel_size, kernel_size]
@@ -51,7 +51,7 @@ def gaussian_blur_image_tensor(
 
 @torch.jit.unused
 def gaussian_blur_image_pil(
-    img: PIL.Image.Image, kernel_size: List[int], sigma: Optional[List[float]] = None
+    img: PIL.Image.Image, kernel_size: Union[int, List[int]], sigma: Union[int, float, List[float], None] = None
 ) -> PIL.Image.Image:
     t_img = pil_to_tensor(img)
     output = gaussian_blur_image_tensor(t_img, kernel_size=kernel_size, sigma=sigma)
@@ -59,8 +59,9 @@ def gaussian_blur_image_pil(
 
 
 def gaussian_blur(
-    inpt: features.InputTypeJIT, kernel_size: List[int], sigma: Optional[List[float]] = None
+    inpt: features.InputTypeJIT, kernel_size: Union[int, List[int]], sigma: Union[int, float, List[float], None] = None
 ) -> features.InputTypeJIT:
+    # TODO: consider deprecating integers on the future
     if isinstance(inpt, torch.Tensor) and (torch.jit.is_scripting() or not isinstance(inpt, features._Feature)):
         return gaussian_blur_image_tensor(inpt, kernel_size=kernel_size, sigma=sigma)
     elif isinstance(inpt, features._Feature):
